W220: the gold set — and the honest finding that it is mostly floor
Five waves of prologue work made this census possible for the first time. Every target passed all
three structural tests simultaneously: **frame size matches retail exactly**, **saved-register
counts match on both banks**, and **it is not a pure register-identity floor**. 849 first-party
functions qualified, 455 with nd <= 300. On paper the closest-to-matching set in the tree, and the
first census where working ASCENDING nd was safe rather than a trap.

Eight agents worked 300-plus of them. Result: nd -149 from seven fixes, **zero matches**, zero
regressions, zero over-window, line endings clean.

  mt_evtMenu 0037abd0   nd88 -> **31**.  An if/else restructured into early returns, eliminating a
                        persistent local and the register pressure it carried.
  h_sfdply 0010c5f0     nd149 -> 110, and 0010cac0 nd192 -> 166 (also 20 bytes smaller).  A loop
                        counter declared `s16` that should be `s32` — the narrow type was costing
                        a sign-extension on every use.
  mdlFile 00320290      nd123 -> 108 and 00344eb0 nd123 -> 118, both `u64` contracts that are
                        really 32-bit, removing `dsll32`/`dsra32`.
  mt_evtMain 00366210   nd108 -> 103, two stack locals whose slots were transposed.
  datCalc 00306bc0      nd178 -> 176, two more inverted assert predicates.

**The finding is the negative, and it is a strong one.** This was the best-filtered target list the
campaign has ever produced, and it still came back overwhelmingly floor. The residuals cluster into
five families, reported independently by every slice: commutative operand orientation; call-argument
setup order before a JAL; independent adjacent load or store reordering; register colouring where
the save counts already agree; and branch/tail-merge layout where all cases jump to one shared
epilogue instead of retail's paired per-group ones. One slice measured all 41 of its targets and
found no fixable defect at all.

That is worth stating plainly: **passing all three structural tests is necessary but nowhere near
sufficient.** The three tests successfully remove the noise, and what is left underneath in the low-nd
band is genuinely the compiler. The productive work is not here.

**A match was reverted on principle.** `btlAction 00297480` reached nd 0 by making an ordinary game
field a `volatile` load to force retail's ordering of two adjacent loads. That is compiler-steering,
not a device access — it tells a future reader the field can change underneath them, which is false,
and the project lints it as an error. Two adjacent independent loads in the opposite order is a
documented floor, and reporting it as one is the correct outcome. Every agent was re-briefed
mid-wave and the rest returned clean floor verdicts rather than reaching for the same trick.

Also newly documented as a floor family: **GP-relative versus absolute addressing that resists a
caller-specific alias.** At `fclCombineMisc 003d6ae0` and `0040a6d0` retail forms an address with
HI16+LO16 while we emit LO16+GPREL16, and adding an `_abs` alias did not move it because MWCC still
chooses GP-relative for an array-of-u32. Previous waves matched three functions on this axis, so it
is per-site and sometimes reachable — but not always, and now we know the shape that resists.

+++ src/Graphics/Model/mdlFile.c
@@ -810,7 +810,7 @@ u8 * FUN_0031fd00(void);
 void FUN_0031fde0(void);
 void FUN_00320080(int param_4,float param_1,int param_5,float param_2,float param_3);
 u32 FUN_00320230(void);
-void FUN_00320290(u64 param_1);
+void FUN_00320290(int param_1);
 void FUN_00320380(float param_1,int *param_2);
 void FUN_003204a0(int *param_4,float param_1,float param_2,float param_3);
 void FUN_00320640(u32 param_1,u32 param_2,int *param_3);
@@ -1268,7 +1268,7 @@ void FUN_00343d10(int param_1);
 void FUN_00343d60(int param_1);
 void FUN_00344720(u64 param_1);
 void FUN_00344b70(int param_1,int param_2);
-u64 FUN_00344eb0(u32 param_1,int param_2);
+u32 FUN_00344eb0(u32 param_1,int param_2);
 u32 FUN_00344fc0(int param_1);
 void FUN_00345080(int param_1);
 void FUN_003450d0(int param_1);
@@ -4530,7 +4530,7 @@ u32 FUN_00320230(void)
 // FUN_00320290 NONMATCHING
 
 
-void FUN_00320290(u64 param_1)
+void FUN_00320290(int param_1)
 
 
 
@@ -38287,7 +38287,7 @@ void FUN_00344b70(int param_1,int param_2)
 // FUN_00344EB0 NONMATCHING
 
 
-u64 FUN_00344eb0(u32 param_1,int param_2)
+u32 FUN_00344eb0(u32 param_1,int param_2)
